feat: Index generation
This effectively implements the workflow I suggested in https://gitlab.com/shosetsuorg/extension-tester/-/merge_requests/4#note_2125100367

Since this means I had to generate a new index (which caused a giantic diff), I have attached the diff according to the tool I wrote in extension-tester.

Since it seemed to be broken, I have also attempted to get the CI script working.

I have also implemented support for GitLab pages and tested it to work from my branch. If you want to migrate to that for hosting, all you'll need to do is add the index to `.gitignore` and remove the if statement from the `validate` CI step. As is, the repo would be hosted at https://shosetsu.gitlab.io/extensions